Tweaking how fused mode works and starting to seperate out sections of main.c into game.c
