+++ src/main.c
@@ -126,227 +126,11 @@ global_variable size_t AUDIO_BUFFER_SIZE = 2048;
 #include "modules/platform.c"
 #include "modules/random.c"
 #include "modules/plugin.c"
-
-
 // Comes last to register modules
 #include "vm.c"
 
-typedef struct {
-  ENGINE* engine;
-  WrenVM* vm;
-  WrenHandle* gameClass;
-  WrenHandle* updateMethod;
-  WrenHandle* drawMethod;
-  double MS_PER_FRAME;
-  double FPS;
-  double lag;
-  uint64_t previousTime;
-  uint64_t currentTime;
-  double elapsed;
-  bool windowBlurred;
-  uint8_t attempts;
-  bool tickRender;
-} LOOP_STATE;
+#include "game.c"
 
-internal void
-LOOP_release(LOOP_STATE* state) {
-  WrenVM* vm = state->vm;
-
-  if (state->drawMethod != NULL) {
-    wrenReleaseHandle(vm, state->drawMethod);
-  }
-
-  if (state->updateMethod != NULL) {
-    wrenReleaseHandle(vm, state->updateMethod);
-  }
-
-  if (state->gameClass != NULL) {
-    wrenReleaseHandle(vm, state->gameClass);
-  }
-}
-
-internal int
-LOOP_processInput(LOOP_STATE* state) {
-  WrenInterpretResult interpreterResult;
-  ENGINE* engine = state->engine;
-  WrenVM* vm = state->vm;
-  engine->mouse.scrollX = 0;
-  engine->mouse.scrollY = 0;
-  SDL_Event event;
-  INPUT_clearText(vm);
-  while(SDL_PollEvent(&event)) {
-    switch (event.type) {
-      case SDL_QUIT:
-        engine->running = false;
-        break;
-      case SDL_WINDOWEVENT:
-        {
-          if (event.window.event == SDL_WINDOWEVENT_RESIZED ||
-              event.window.event == SDL_WINDOWEVENT_SIZE_CHANGED) {
-            SDL_RenderGetViewport(engine->renderer, &(engine->viewport));
-            break;
-          }
-          if (event.window.event == SDL_WINDOWEVENT_FOCUS_LOST) {
-#ifdef __EMSCRIPTEN__
-            AUDIO_ENGINE_pause(engine->audioEngine);
-#endif
-            state->windowBlurred = true;
-          } else if (event.window.event == SDL_WINDOWEVENT_FOCUS_GAINED) {
-#ifdef __EMSCRIPTEN__
-            AUDIO_ENGINE_resume(engine->audioEngine);
-#endif
-            ENGINE_updateTextRegion(engine);
-            state->windowBlurred = false;
-          }
-        } break;
-      case SDL_KEYDOWN:
-      case SDL_KEYUP:
-        {
-          SDL_Keycode keyCode = event.key.keysym.sym;
-          if (keyCode == SDLK_F3 && event.key.state == SDL_PRESSED && event.key.repeat == 0) {
-            engine->debugEnabled = !engine->debugEnabled;
-          } else if (keyCode == SDLK_F2 && event.key.state == SDL_PRESSED && event.key.repeat == 0) {
-            ENGINE_takeScreenshot(engine);
-          } else if (event.key.repeat == 0) {
-            char* buttonName = strToLower((char*)SDL_GetKeyName(keyCode));
-            interpreterResult = INPUT_update(vm, DOME_INPUT_KEYBOARD, buttonName, event.key.state == SDL_PRESSED);
-            free(buttonName);
-            if (interpreterResult != WREN_RESULT_SUCCESS) {
-              return EXIT_FAILURE;
-            }
-          }
-        } break;
-      case SDL_TEXTEDITING:
-        {
-          if (utf8len(event.edit.text) > 0) {
-            INPUT_setCompositionText(vm, event.edit.text, event.edit.start, event.edit.length);
-          }
-        } break;
-      case SDL_TEXTINPUT:
-        {
-          if (utf8len(event.text.text) > 0) {
-            INPUT_addText(vm, event.text.text);
-          }
-        } break;
-
-      case SDL_CONTROLLERDEVICEADDED:
-        {
-          GAMEPAD_eventAdded(vm, event.cdevice.which);
-        } break;
-      case SDL_CONTROLLERDEVICEREMOVED:
-        {
-          GAMEPAD_eventRemoved(vm, event.cdevice.which);
-        } break;
-      case SDL_CONTROLLERBUTTONDOWN:
-      case SDL_CONTROLLERBUTTONUP:
-        {
-          SDL_ControllerButtonEvent cbutton = event.cbutton;
-          const char* buttonName = GAMEPAD_stringFromButton(cbutton.button);
-          interpreterResult = GAMEPAD_eventButtonPressed(vm, cbutton.which, buttonName, cbutton.state == SDL_PRESSED);
-          if (interpreterResult != WREN_RESULT_SUCCESS) {
-            return EXIT_FAILURE;
-          }
-        } break;
-      case SDL_MOUSEWHEEL:
-        {
-          int dir = event.wheel.direction == SDL_MOUSEWHEEL_NORMAL ? 1 : -1;
-          engine->mouse.scrollX += event.wheel.x * dir;
-          // Down should be positive to match the direction of rendering.
-          engine->mouse.scrollY += event.wheel.y * -dir;
-        } break;
-      case SDL_MOUSEBUTTONDOWN:
-      case SDL_MOUSEBUTTONUP:
-        {
-          char* buttonName;
-          switch (event.button.button) {
-            case SDL_BUTTON_LEFT: buttonName = "left"; break;
-            case SDL_BUTTON_MIDDLE: buttonName = "middle"; break;
-            case SDL_BUTTON_RIGHT: buttonName = "right"; break;
-            case SDL_BUTTON_X1: buttonName = "x1"; break;
-            default:
-            case SDL_BUTTON_X2: buttonName = "x2"; break;
-          }
-          bool state = event.button.state == SDL_PRESSED;
-          interpreterResult = INPUT_update(vm, DOME_INPUT_MOUSE, buttonName, state);
-          if (interpreterResult != WREN_RESULT_SUCCESS) {
-            return EXIT_FAILURE;
-          }
-        } break;
-      case SDL_USEREVENT:
-        {
-          ENGINE_printLog(engine, "Event code %i\n", event.user.code);
-          if (event.user.code == EVENT_LOAD_FILE) {
-            FILESYSTEM_loadEventComplete(&event);
-          }
-        }
-    }
-  }
-  if (inputCaptured) {
-    interpreterResult = INPUT_commit(vm);
-    if (interpreterResult != WREN_RESULT_SUCCESS) {
-      return EXIT_FAILURE;
-    }
-  }
-  return EXIT_SUCCESS;
-}
-
-internal int
-LOOP_render(LOOP_STATE* state) {
-  if (PLUGIN_COLLECTION_runHook(state->engine, DOME_PLUGIN_HOOK_PRE_DRAW) != DOME_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  };
-  WrenInterpretResult interpreterResult;
-  wrenEnsureSlots(state->vm, 8);
-  wrenSetSlotHandle(state->vm, 0, state->gameClass);
-  wrenSetSlotDouble(state->vm, 1, ((double)state->lag / state->MS_PER_FRAME));
-  interpreterResult = wrenCall(state->vm, state->drawMethod);
-  if (interpreterResult != WREN_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  }
-  if (PLUGIN_COLLECTION_runHook(state->engine, DOME_PLUGIN_HOOK_POST_DRAW) != DOME_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  };
-
-
-  return EXIT_SUCCESS;
-}
-
-internal void
-LOOP_flip(LOOP_STATE* state) {
-
-  state->engine->debug.elapsed = state->elapsed;
-  if (state->engine->debugEnabled) {
-    ENGINE_drawDebug(state->engine);
-  }
-  // Flip Buffer to Screen
-  SDL_UpdateTexture(state->engine->texture, 0, state->engine->canvas.pixels, state->engine->canvas.width * 4);
-  // clear screen
-  SDL_RenderClear(state->engine->renderer);
-  SDL_RenderCopy(state->engine->renderer, state->engine->texture, NULL, NULL);
-  SDL_RenderPresent(state->engine->renderer);
-}
-
-internal int
-LOOP_update(LOOP_STATE* state) {
-  WrenInterpretResult interpreterResult;
-
-  if (PLUGIN_COLLECTION_runHook(state->engine, DOME_PLUGIN_HOOK_PRE_UPDATE) != DOME_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  };
-
-  wrenEnsureSlots(state->vm, 8);
-  wrenSetSlotHandle(state->vm, 0, state->gameClass);
-  interpreterResult = wrenCall(state->vm, state->updateMethod);
-  if (interpreterResult != WREN_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  }
-  if (PLUGIN_COLLECTION_runHook(state->engine, DOME_PLUGIN_HOOK_POST_UPDATE) != DOME_RESULT_SUCCESS) {
-    return EXIT_FAILURE;
-  };
-  // updateAudio()
-  AUDIO_ENGINE_update(state->engine->audioEngine, state->vm);
-  return EXIT_SUCCESS;
-}
 
 internal void
 printTitle(ENGINE* engine) {
@@ -392,33 +176,6 @@ printUsage(ENGINE* engine) {
   ENGINE_printLog(engine, "  -v --version        Show version.\n");
 }
 
-void DOME_loop(void* data) {
-  LOOP_STATE loop = *((LOOP_STATE*)data);
-  loop.currentTime = SDL_GetPerformanceCounter();
-  loop.elapsed = 1000 * (loop.currentTime - loop.previousTime) / (double) SDL_GetPerformanceFrequency();
-  loop.previousTime = loop.currentTime;
-  loop.lag += loop.elapsed;
-
-  if (loop.lag >= loop.MS_PER_FRAME) {
-    LOOP_processInput(&loop);
-    if (loop.windowBlurred) {
-      loop.lag = 0;
-      loop.tickRender = true;
-      return;
-    }
-    LOOP_update(&loop);
-    if (loop.tickRender) {
-      LOOP_render(&loop);
-      LOOP_flip(&loop);
-    }
-    loop.tickRender = !loop.tickRender;
-    loop.lag = mid(0, loop.lag - loop.MS_PER_FRAME, loop.MS_PER_FRAME);
-  }
-  *((LOOP_STATE*)data) = loop;
-}
-
-
-
 int main(int argc, char* args[])
 {
   // configuring the buffer has to be first
@@ -444,6 +201,39 @@ int main(int argc, char* args[])
   loop.engine = &engine;
 
 #ifndef __EMSCRIPTEN__
+  // Is this web? If No...
+    // Are we running in fused mode?
+    // Did we get a path?
+  // No - Try opening game.egg, or main.wren
+  // Yes - Try opening it. If it's an egg, run it, otherwise don't.
+
+  char* binaryPath = FUSE_getExecutablePath();
+  if (binaryPath != NULL) {
+    // Check if end of file has marker
+    FILE* self = fopen(binaryPath, "rb");
+    int result = fseek (self, -((long int)sizeof(DOME_FUSED_HEADER)), SEEK_END);
+    if (result == 0) {
+      DOME_FUSED_HEADER header;
+      result = fread(&header, sizeof(DOME_FUSED_HEADER), 1, self);
+      if (result == 1) {
+        if (memcmp("DOME", header.magic1, 4) == 0 && memcmp("DOME", header.magic2, 4) == 0) {
+          if (header.version == 1) {
+            engine.tar = malloc(sizeof(mtar_t));
+            FUSE_open(engine.tar, self, header.offset);
+  //          engine.argv[1] = NULL;
+            engine.fused = true;
+          } else {
+            printf("DOME is in fused mode, but the data is corrupt.");
+            fclose(self);
+          }
+        }
+      }
+    } else {
+      fclose(self);
+    }
+  }
+  free(binaryPath);
+
   struct optparse_long longopts[] = {
     {"help", 'h', OPTPARSE_NONE},
     {"version", 'v', OPTPARSE_NONE},
@@ -460,63 +250,65 @@ int main(int argc, char* args[])
   struct optparse options;
   optparse_init(&options, args);
   options.permute = 0;
-  if (argc > 1) {
-    while ((option = optparse_long(&options, longopts, NULL)) != -1) {
-      switch (option) {
-        case 'b':
-          {
-            int shift = atoi(options.optarg);
-            if (shift == 0) {
-              // If it wasn't valid, set to a meaningful default.
-              AUDIO_BUFFER_SIZE = 2048;
-            }
-            AUDIO_BUFFER_SIZE = 1 << shift;
-          } break;
-#ifdef __MINGW32__
-        case 'c': {
-                    AllocConsole();
-                    freopen("CONIN$", "r", stdin);
-                    freopen("CONOUT$", "w", stdout);
-                    freopen("CONOUT$", "w", stderr);
-                  } break;
-#endif
-        case 'd':
-                  DEBUG_MODE = true;
-                  ENGINE_printLog(&engine, "Debug Mode enabled\n");
-                  break;
-        case 'h':
-                  printTitle(&engine);
-                  printUsage(&engine);
-                  goto cleanup;
-        case 'v':
-                  printTitle(&engine);
-                  printVersion(&engine);
-                  goto cleanup;
-        case '?':
-                  fprintf(stderr, "%s: %s\n", args[0], options.errmsg);
-                  result = EXIT_FAILURE;
-                  goto cleanup;
-      }
-    }
 
-    static const struct {
-      char name[8];
-      char letter;
-      int (*cmd)(ENGINE*, char **);
-    } cmds[] = {
-      {"fuse", 'f',  FUSE_perform },
-      {"embed", 'e', EMBED_perform },
-      {"help",  'h', HELP_perform },
-      {"nest", 'n', NEST_perform }
-    };
-    int ncmds = sizeof(cmds) / sizeof(*cmds);
-    subargv = args + options.optind;
-    // If we match a subcommand, execute it
-    for (int i = 0; i < ncmds; i++) {
-      if (!strncmp(cmds[i].name, subargv[0], strlen(subargv[0]))) { // || subargv[0][0] == cmds[i].letter) {
-      // if (!strcmp(cmds[i].name, subargv[0])) { // || subargv[0][0] == cmds[i].letter) {
-        result = cmds[i].cmd(&engine, subargv);
-        goto cleanup;
+  if (!engine.fused) {
+    if (argc > 1) {
+      while ((option = optparse_long(&options, longopts, NULL)) != -1) {
+        switch (option) {
+          case 'b':
+            {
+              int shift = atoi(options.optarg);
+              if (shift == 0) {
+                // If it wasn't valid, set to a meaningful default.
+                AUDIO_BUFFER_SIZE = 2048;
+              }
+              AUDIO_BUFFER_SIZE = 1 << shift;
+            } break;
+#ifdef __MINGW32__
+          case 'c': {
+                      AllocConsole();
+                      freopen("CONIN$", "r", stdin);
+                      freopen("CONOUT$", "w", stdout);
+                      freopen("CONOUT$", "w", stderr);
+                    } break;
+#endif
+          case 'd':
+                    DEBUG_MODE = true;
+                    ENGINE_printLog(&engine, "Debug Mode enabled\n");
+                    break;
+          case 'h':
+                    printTitle(&engine);
+                    printUsage(&engine);
+                    goto cleanup;
+          case 'v':
+                    printTitle(&engine);
+                    printVersion(&engine);
+                    goto cleanup;
+          case '?':
+                    fprintf(stderr, "%s: %s\n", args[0], options.errmsg);
+                    result = EXIT_FAILURE;
+                    goto cleanup;
+        }
+      }
+
+      static const struct {
+        char name[8];
+        char letter;
+        int (*cmd)(ENGINE*, char **);
+      } cmds[] = {
+        {"fuse", 'f',  FUSE_perform },
+        {"embed", 'e', EMBED_perform },
+        {"help",  'h', HELP_perform },
+        {"nest", 'n', NEST_perform }
+      };
+      int ncmds = sizeof(cmds) / sizeof(*cmds);
+      subargv = args + options.optind;
+      // If we match a subcommand, execute it
+      for (int i = 0; i < ncmds; i++) {
+        if (!strncmp(cmds[i].name, subargv[0], strlen(subargv[0]))) {
+          result = cmds[i].cmd(&engine, subargv);
+          goto cleanup;
+        }
       }
     }
   }
@@ -525,13 +317,6 @@ int main(int argc, char* args[])
 
   // assume we are trying to play the game
   {
-    char* defaultEggName = "game.egg";
-    char* mainFileName = "main.wren";
-
-    char* base = BASEPATH_get();
-
-    char pathBuf[PATH_MAX];
-    char* fileName = NULL;
 
 #ifndef __EMSCRIPTEN__
     // Get non-option args list
@@ -555,6 +340,7 @@ int main(int argc, char* args[])
     if (domeArgCount > 1) {
       arg = engine.argv[1];
     }
+
 #else
     engine.argv = calloc(2, sizeof(char*));
     engine.argv[0] = args[0];
@@ -564,72 +350,52 @@ int main(int argc, char* args[])
     char* arg = NULL;
 #endif
 
+    char* defaultEggName = "game.egg";
+    char* mainFileName = "main.wren";
+    char pathBuf[PATH_MAX];
+    char* fileName = NULL;
+    char* base = BASEPATH_get();
 
-    // Get establish the path components: filename(?) and basepath.
-    if (arg != NULL) {
-      strcpy(pathBuf, base);
-      strcat(pathBuf, arg);
-      if (isDirectory(pathBuf)) {
-        BASEPATH_set(pathBuf);
-        autoResolve = true;
-      } else {
-        char* dirc = strdup(pathBuf);
-        char* basec = strdup(pathBuf);
-        // This sets the filename used.
-        fileName = strdup(basename(dirc));
-        BASEPATH_set(dirname(basec));
-        free(dirc);
-        free(basec);
-      }
-
-      base = BASEPATH_get();
-    }
-
-    // If a filename is given in the path, use it, or assume its 'game.egg'
-    strcpy(pathBuf, base);
-    strcat(pathBuf, !autoResolve ? fileName : defaultEggName);
-    chdir(base);
-
-    if (doesFileExist(pathBuf)) {
-      // the current path exists, let's see if it's a TAR file.
-      engine.tar = malloc(sizeof(mtar_t));
-      int tarResult = mtar_open(engine.tar, pathBuf, "r");
-      if (tarResult == MTAR_ESUCCESS) {
-        ENGINE_printLog(&engine, "Loading bundle %s\n", pathBuf);
-        engine.argv[1] = strdup(pathBuf);
-      } else {
-        // Not a valid tar file.
-        free(engine.tar);
-        engine.tar = NULL;
-      }
-    } else {
-#ifndef __EMSCRIPTEN__
-      char* binaryPath = FUSE_getExecutablePath();
-      if (binaryPath != NULL) {
-        // Check if end of file has marker
-        FILE* self = fopen(binaryPath, "rb");
-        int result = fseek (self, -((long int)sizeof(DOME_FUSED_HEADER)), SEEK_END);
-        if (result == 0) {
-          DOME_FUSED_HEADER header;
-          result = fread(&header, sizeof(DOME_FUSED_HEADER), 1, self);
-          if (result == 1) {
-            if (memcmp("DOME", header.magic1, 4) == 0 && memcmp("DOME", header.magic2, 4) == 0) {
-              if (header.version == 1) {
-                engine.tar = malloc(sizeof(mtar_t));
-                FUSE_open(engine.tar, self, header.offset);
-                engine.argv[1] = NULL;
-              } else {
-                printf("DOME is in fused mode, but the data is corrupt.");
-                fclose(self);
-              }
-            }
-          }
+    if (!engine.fused) {
+      // Get establish the path components: filename(?) and basepath.
+      if (arg != NULL) {
+        strcpy(pathBuf, base);
+        strcat(pathBuf, arg);
+        if (isDirectory(pathBuf)) {
+          BASEPATH_set(pathBuf);
+          autoResolve = true;
         } else {
-          fclose(self);
+          char* dirc = strdup(pathBuf);
+          char* basec = strdup(pathBuf);
+          // This sets the filename used.
+          fileName = strdup(basename(dirc));
+          BASEPATH_set(dirname(basec));
+          free(dirc);
+          free(basec);
+        }
+
+        base = BASEPATH_get();
+      }
+
+      // If a filename is given in the path, use it, or assume its 'game.egg'
+      strcpy(pathBuf, base);
+      strcat(pathBuf, !autoResolve ? fileName : defaultEggName);
+      chdir(base);
+
+      if (doesFileExist(pathBuf)) {
+        // the current path exists, let's see if it's a TAR file.
+        engine.tar = malloc(sizeof(mtar_t));
+        int tarResult = mtar_open(engine.tar, pathBuf, "r");
+        if (tarResult == MTAR_ESUCCESS) {
+          ENGINE_printLog(&engine, "Loading bundle %s\n", pathBuf);
+          engine.argv[1] = strdup(pathBuf);
+        } else {
+          // Not a valid tar file.
+          free(engine.tar);
+          engine.tar = NULL;
         }
       }
-      free(binaryPath);
-#endif
+
     }
 
     if (engine.tar != NULL) {
@@ -709,7 +475,7 @@ int main(int argc, char* args[])
   SDL_ShowWindow(engine.window);
 
   loop.lag = loop.MS_PER_FRAME;
-  result = LOOP_processInput(&loop);
+  result = DOME_processInput(&loop);
   if (result != EXIT_SUCCESS) {
     goto vm_cleanup;
   }
@@ -722,7 +488,7 @@ int main(int argc, char* args[])
 
     // processInput()
     if (loop.windowBlurred) {
-      result = LOOP_processInput(&loop);
+      result = DOME_processInput(&loop);
       if (result != EXIT_SUCCESS) {
         goto vm_cleanup;
       }
@@ -753,46 +519,46 @@ int main(int argc, char* args[])
 
     if (engine.lockstep) {
       if (loop.lag >= loop.MS_PER_FRAME) {
-        result = LOOP_processInput(&loop);
+        result = DOME_processInput(&loop);
         if (result != EXIT_SUCCESS) {
           goto vm_cleanup;
         }
-        result = LOOP_update(&loop);
+        result = DOME_update(&loop);
         if (result != EXIT_SUCCESS) {
           goto vm_cleanup;
         }
-        result = LOOP_render(&loop);
+        result = DOME_render(&loop);
         if (result != EXIT_SUCCESS) {
           goto vm_cleanup;
         }
         loop.lag = mid(0, loop.lag - loop.MS_PER_FRAME, loop.MS_PER_FRAME);
-        LOOP_flip(&loop);
+        DOME_flip(&loop);
       }
     } else {
       loop.attempts = 5;
       while (loop.attempts > 0 && loop.lag >= loop.MS_PER_FRAME) {
         loop.attempts--;
 
-        result = LOOP_processInput(&loop);
+        result = DOME_processInput(&loop);
         if (result != EXIT_SUCCESS) {
           goto vm_cleanup;
         }
         // update()
-        result = LOOP_update(&loop);
+        result = DOME_update(&loop);
         if (result != EXIT_SUCCESS) {
           goto vm_cleanup;
         }
         loop.lag -= loop.MS_PER_FRAME;
       }
       // render();
-      result = LOOP_render(&loop);
+      result = DOME_render(&loop);
       if (result != EXIT_SUCCESS) {
         goto vm_cleanup;
       }
       if (loop.attempts == 0) {
         loop.lag = 0;
       }
-      LOOP_flip(&loop);
+      DOME_flip(&loop);
     }
 
 
@@ -820,7 +586,7 @@ vm_cleanup:
     wrenReleaseHandle(vm, initMethod);
   }
 
-  LOOP_release(&loop);
+  DOME_release(&loop);
 
   if (bufferClass != NULL) {
     wrenReleaseHandle(vm, bufferClass);
